requirements.txt
tensorflow==2.3.1
requests==2.22
numpy==1.17.0
Pillow==6.1.0